Fix warning

+++ src/frontend/cxx-cexpr.c
@@ -148,7 +148,7 @@ const_value_t* const_value_get_integer(cvalue_uint_t value, int num_bytes, char 
             || num_bytes < 0, "Invalid num_bytes = %d\n", num_bytes);
 
     if (!sign
-            && (num_bytes < sizeof(value)))
+            && (num_bytes < (int)sizeof(value)))
     {
         // Make sure higher bits are set to zero if this value is unsigned
         cvalue_uint_t mask = ~(cvalue_uint_t)0;
